requirements.txt
astor==0.8.1
Babel==2.9.1
backports.entry-points-selectable==1.1.0
bce-python-sdk==0.8.62
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
cfgv==3.3.1
click==8.0.3
cssselect==1.1.0
cssutils==2.3.0
cycler==0.10.0
distlib==0.3.3
et-xmlfile==1.1.0
filelock==3.3.1
flake8==4.0.1
Flask==2.0.2
Flask-Babel==2.0.0
future==0.18.2
gast==0.3.3
identify==2.3.1
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.2.0
iopath==0.1.9
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
lmdb==1.2.1
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
mkl-fft==1.3.1
mkl-service==2.4.0
networkx==2.6.3
nodeenv==1.6.0
olefile==0.46
opencv-python==4.2.0.32
openpyxl==3.0.9
paddlepaddle-gpu==2.2.0.post101
pandas==1.3.4
Pillow==8.4.0
platformdirs==2.4.0
Polygon3==3.0.9.1
portalocker==2.3.2
pre-commit==2.15.0
premailer==3.10.0
protobuf==3.17.2
pyclipper==1.3.0
pycodestyle==2.8.0
pycryptodome==3.11.0
pyflakes==2.4.0
pyparsing==3.0.1
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
scikit-image==0.17.2
scipy==1.7.1
Shapely==1.8.0
shellcheck-py==0.7.2.1
tifffile==2021.10.12
toml==0.10.2
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.9.0
visualdl==2.2.1
Werkzeug==2.0.2
zipp==3.6.0
